+++ tiptapy/image.py
@@ -1,7 +1,6 @@
 # Image file type and it's MIME type mappings that are suported by tiptapy.
 # Detailed documentation can be found about Image file type and format guide.
 # Link: https://developer.mozilla.org/en-US/docs/Web/Media/Formats/Image_types
-from html import escape
 from os.path import splitext
 
 
@@ -24,4 +23,4 @@ SUPPORTED_FORMATS_MAP = SupportedFormatsMapper(
 def url2mime(url):
     ext = splitext(url)[-1]
     ext = (ext[1:] if ext.startswith(".") else ext).upper()
-    return escape(SUPPORTED_FORMATS_MAP[ext])
+    return SUPPORTED_FORMATS_MAP[ext]